requirements.txt
# Discord bot dependencies
discord.py>=2.0.0
python-dotenv>=0.19.0

# Spotify integration dependencies
spotipy>=2.23.0
requests>=2.28.0
aiohttp>=3.8.0